+++ examples/mock_app_debug.py
@@ -4,15 +4,15 @@ from time import sleep
 dot = MockBlueDot(auto_start_server = False)
 
 def pressed(pos):
-    print("Pressed: x-{} y-{} angle-{} distance-{} middle-{} top-{} bottom-{} left-{} right-{}".format(pos.x, pos.y, pos.angle, pos.distance, pos.middle, pos.top, pos.bottom, pos.left, pos.right))
+    print("Pressed: x={} y={} angle={} distance={} middle={} top={} bottom={} left={} right={} time={}".format(pos.x, pos.y, pos.angle, pos.distance, pos.middle, pos.top, pos.bottom, pos.left, pos.right, time()))
 
 def released():
-    print("Released: x-{} y-{}".format(dot.position.x, dot.position.y))
+    print("Released: x={} y={}".format(dot.position.x, dot.position.y))
     print()
 
 def moved(pos):
-    print("Moved: x-{} y-{}".format(pos.x, pos.y))
-
+    print("Moved: x={} y={}".format(pos.x, pos.y))
+    
 dot.when_pressed = pressed
 dot.when_released = released
 dot.when_moved = moved